+++ src/pages/login/login.tsx
@@ -89,7 +89,7 @@ export function Login() {
       const request = await auth.authenticate(login, senha, loginType);
       if (request.token) {
         setLoading(false);
-        navigate("/instrucoes");
+        navigate("/curso");
       } else {
         toast.error("Não foi possível entrar, verifique as credenciais!");
         setLoading(false);
